Add database migrations

+++ routes/services.js
@@ -1,24 +1,31 @@
 const jwt = require('jsonwebtoken')
-
-// TODO In each route check if req.user.type==='service'
+const mongodb = require('mongodb')
 
 module.exports = ({db, app, _PRIVATE_KEY}) => {
 
     app.delete('/services/my', function (req, res) {
-        db.collection('services').deleteOne({_id: new mongodb.ObjectID(req.user.id)});
-        return res.send('ok')
+        if (req.user.type==='service') {
+            db.collection('services').deleteOne({_id: new mongodb.ObjectID(req.user.id)});
+            return res.send('ok')
+        } else {
+            return res.send('You are not a service')
+        }
     })
 
     app.put('/services/my', function (req, res) {
-        if (req.body.name) {
-            db.collection('services').updateOne({_id: new mongodb.ObjectID(req.user.id)}, { 
-                $set: {
-                    pwd: req.body.pwd
-                }
-            })
-            return res.send('ok')
+        if (req.user.type==='service') {
+            if (req.body.pwd) {
+                db.collection('services').updateOne({_id: new mongodb.ObjectID(req.user.id)}, { 
+                    $set: {
+                        pwd: req.body.pwd
+                    }
+                })
+                return res.send('ok')
+            } else {
+                return res.send('err: passami il parametro pwd nel body')
+            }
         } else {
-            return res.send('err: passami il parametro name nel body')
+            return res.send('You are not a service')
         }
     })
